Decompose spend contract tests
